feat: config params

+++ Config.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 #include <string>
 
 namespace Config {
@@ -8,11 +9,11 @@ namespace Config {
     // Config values 
     inline int         numCpu            = 4;
     inline std::string scheduler         = "fcfs";
-    inline int         quantumCycles     = 5;
-    inline int         batchProcessFreq  = 1;
-    inline int         minIns            = 1000;
-    inline int         maxIns            = 2000;
-    inline int         delayPerExec      = 0;
+    inline uint64_t    quantumCycles     = 5;
+    inline uint64_t    batchProcessFreq  = 1;
+    inline uint64_t    minIns            = 1000;
+    inline uint64_t    maxIns            = 2000;
+    inline uint64_t    delayPerExec      = 0;
 
     // Memory manager parameters (first-fit flat allocator).
     inline int         maxOverallMem     = 16384;  // total bytes of main memory
@@ -22,8 +23,8 @@ namespace Config {
 
     // Legacy aliases so existing Scheduler/Process code compiles unchanged
     inline int& NUM_CORES             = numCpu;
-    inline int& PRINTS_PER_PROCESS    = maxIns;   // temporary; replace later
-    inline int& PER_INSTRUCTION_DELAY_MS = delayPerExec;
+    inline uint64_t& PRINTS_PER_PROCESS = maxIns;   // temporary; replace later
+    inline uint64_t& PER_INSTRUCTION_DELAY_MS = delayPerExec;
     inline bool WRITE_PRINT_FILES     = true;
 
     // Reads config.txt from the working directory.
